Change how images are added to properties:
The default behavior when adding an image URL is to download and embed the image in the document.
An instance of ICesiumUrlResolver can be passed to control how the URL should be embedded.

Rewrite most of the image encoding functionality, and remove the #if !CSToJava so it works in
Java as well.