Change image property writing API again.
Make the choice of Embed vs Link explicit via a CesiumResourceBehavior enumeration.
Also change ImageCesiumWriter to have an ICesiumValuePropertyWriter<CesiumResource>
instead of string, which is a new class containing the URL along with the choice of how
to reference the reference.

CesiumFormattingHelper.DownloadUrlIntoDataUri should not check relative vs absolute URLs.
Any URLs that are not retriveable should either be absolutized externally by client code, or
added as LinkTo, not Embed.