Add pom.xml for Maven
Added a pom.xml so that maven will automagically download dependencies and this
artifact can be uploaded to maven central. I changed the default source, test
and resource paths to their current location in the source tree.

For the build to be portable cs2j.jar still has to be uploaded to maven central.