Make materials more closely match Cesium
Grid and Image materials split out Cartesian2 properties into multiple individual properties.  In practice this is not very useful and can bloat the size of the CZML if both are chaning.  This change also makes it easier to implement in Cesium and gets rid of some workarounds we were currently using.