Fix GridMaterial and ImageMaterial
#73 used `Cartesian2Value.jsonschema` which is incorrect.
I was tempted to create a `Cartesian2.jsonschema`, but the practice we
have been using is to create property-specific writers, i.e. Repeat.jsonschena,
LineThickness.jsonschema, etc.. so that is what I did here.

I think we need to revisit this strategy (and I might do that ASAP) but I wanted to get everything working properly before refactoring.